Group add/remove: do not report success until connect
It turns out our _tests_ had a race that was very unlikely in production
code:

1. Host A registers with group G
2. Host B registers with group G
3. Host A connects to RabbitMQ
4. Host A sends message M to group G
5. Host B connects to RabbitMQ

In this race, Host B won't receive M -- because it's not connected and so
not registered with G.

That's by design.

But what happens if we add something else!

1. Host A registers with group G
2. Host B registers with group G
**2.5. Somehow (_not_ RabbitMQ), Host B tells host A it's in group G
3. Host A connects to RabbitMQ
4. Host A sends message M to group G
5. Host B connects to RabbitMQ

Now, Host A would _expect_ its message M to reach host B in most cases.

The added complexity is a side channel: A and B communicating their
expectations _outside_ Channels. That's not common ... but we might as
well support that case.

It turns out we had a unit test that occasionally failed here. Its
side-channel: the test opened two connections and assumed they were
connected when `group_add()` returned.

There's also a new _groups_lock added. It takes less effort to add the
lock than to explain why it isn't necessary. And since `group_discard()`
reads `self._channel`, it actually _might_ be necessary.