Make race in test less likely.
Blitzing messages to the server meant the server would sometimes nack
because the client hadn't caught up. The behavior (nacking message 3
instead of message 4) is the behavior we want in that case ... but that
isn't what we're testing. A couple of asyncio.sleep() calls should make
this race infinitessimally unlikely.