Expire messages after receiving them from RabbitMQ
Previously, a message sent to a nonexistent channel would be queued
forever. After enough such messages, the entire process would stop
delivering messages (because local_capacity would be full).

The culprit in Workbench's case was a `group_add()` without a matching
`group_discard()`. The `group_discard()` was missing because Daphne
cancels `disconnect()` on a Websockets consumer if it takes >10s.

... but this is a more general problem in the Channels specification.
Consumers don't register themselves, so messages are _usually_ sent to
nonexistent consumers.