Ensure RabbitMQ queue does not exist before declaring
At Workbench, we ran into an error in production: during reconnect
to a cluster, we'd get an error during _setup_subscriptions() indicating
that the queue was not declared. But the queue is declared in the
previous function call, _setup_connection()! I'm guessing this is a race
involving cluster state -- the queue name might not have been
garbage-collected from the new RabbitMQ node the client reconnected to.

If I'm right, this will fix a race when reconnecting to a different
cluster node.